aceit_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.127.0
h11==0.16.0
idna==3.11
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
pypdfium2==5.2.0
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25.0
google-generativeai==0.3.2
gTTS
openai-whisper
